refactor #66: Page 컴포넌트 분리
