+++ fe/src/Pages/Result.tsx
